반차까지 됩니다여

+++ src/main/java/com/mycompany/webapp/vacation/model/Vacation.java
@@ -15,7 +15,7 @@ public class Vacation {
 	private String dates;
 	private int vacationCategoryId;
 	private String empId;
-	private int countDay;
+	private double countDay;
 
 	private String approvalEmpId;
